chore: Fix typo seperator->separator
